kernel-*headers: Symlink videodev2 headers to linux/
The videodev2*.h headers were all stored exclusively in:
  kernel-*-headers/kernel-headers/

This is a huge problem as in some places, these headers are included as
"<videodev2.h>" and in other places "<linux/videodev2.h>". In the latter
case, the headers are completely missed and unused. At least somehow now
on Exynos 7904 with a 4.4 kernel, where until now, that was causing no
apparent issues elsewhere.

On 7904, the proper headers going unused was causing incorrect V4L2
ioctls in libhwjpeg and libacryl. So, fix that by symlinking the
videodev2 headers to kernel-*-headers/kernel-headers/linux/, for 4.4 and
also any newer kernels.

Change-Id: Id9aedba21e4a68d20fc83f6e49b367b686e19e18